+++ src/main/java/com/mfelipeee/betterarrows/BetterArrows.java
@@ -94,16 +94,9 @@ public class BetterArrows {
     }
 
     private void commonSetup(FMLCommonSetupEvent event) {
-        // Some common setup code
         LOGGER.info("HELLO FROM COMMON SETUP");
-
-        if (Config.LOG_DIRT_BLOCK.getAsBoolean()) {
-            LOGGER.info("DIRT BLOCK >> {}", BuiltInRegistries.BLOCK.getKey(Blocks.DIRT));
-        }
-
-        LOGGER.info("{}{}", Config.MAGIC_NUMBER_INTRODUCTION.get(), Config.MAGIC_NUMBER.getAsInt());
-
-        Config.ITEM_STRINGS.get().forEach((item) -> LOGGER.info("ITEM >> {}", item));
+        LOGGER.info("Material sounds enabled >> {}", Config.ENABLE_MATERIAL_SOUNDS.getAsBoolean());
+        LOGGER.info("Water slowdown enabled >> {}", Config.ENABLE_WATER_SLOWDOWN.getAsBoolean());
     }
 
     // Add the example block item to the building blocks tab